manager: accept internal reactivation if the master changed
When a master with autoconnect-slaves=yes is activated, it activates
slaves too. Since the master is activating, it means it was
disconnected before and thus its slaves were not active as well.

However, it could happen that the slave device had another connection
active with a different (or no) master. In such case the check in
_internal_activation_auth_done() prevents the activation of the slave
connection.

The patch addresses this specific case by checking whether the master
of a connection changed since activation and allowing the internal
activation to proceed in such case. It restores the NM 1.8 behavior of
autoconnecting a slave even if the connection is already active
(which, as explained above, can happen only when the master property
of a slave gets changed).

https://bugzilla.redhat.com/show_bug.cgi?id=1548265