device: drop capture_lease_config() during connection assumption
Drop capture_lease_config(). It was added by commit
0321073b3cb9060817c7743e1c203c6b72659dfe.

Note that it was only called by

  - capture_lease_config()
    - update_ext_ip_config(addr_family=AF_INET, initial=TRUE)
      - nm_device_capture_initial_config()
        - manater:get_existing_connection()
          - recheck_assume_connection()

It had only effect when the device had IPv4 permanent addresses.
But then, capture_lease_config() would go on and iterate over
all connections (sorted by last-connect timestamp). It would
consider connection candidates that are compatible with the device,
and try to read the lease information from disk

It's really unclear what this means. For assuming (graceful take over),
do we need the lease information in the device? I don't think so,
because we will match an existing connection. The lease information
shall be read while activating (if necessary).

For external connections, we don't even have a matching connection
and we always generate a new one. It doesn't seem right to consider
leases from disk, for a different connection.

Just drop this. It's really ugly. If this causes an issue, it must be
fixed differently. We want to behave determinstically and well defined.
I don't even comprehend all the implications of what this had.

Also note that update_ext_ip_config() no longer clears
priv->dev_ip4_config.