device: rework initial capturing of device config to not capture resolve-conf
This was called by via
  - update_ext_ip_config(initial=TRUE)
    - nm_device_capture_initial_config()
      - manater:get_existing_connection()
        - recheck_assume_connection()
and would parse resolv.conf, and try to fill the device IP config
with nameservers and dns-options.

But why? It would only have effect if NM was started with
nm_dns_manager_get_resolv_conf_explicit(), but is that really sensible?
And it would only take effect on devices that have a default route.
And for what is this information even used?

Let's not do it this way. If we need this information for assuming or
external sys-iface mode, then it should be explicitly loaded at the
appropriate moment.

For now, drop it and see what breaks. Then we can fix it properly. If
it even matters.